review remarks

+++ lib/quantization/src/turboquant/simd/query2bit/x64.rs
@@ -397,8 +397,7 @@ pub unsafe fn score_2bit_internal_avx2(a: &[u8], b: &[u8]) -> f32 {
 // ------------------------------------------------------------------
 // score_2bit_internal_weighted — TQ+ symmetric path with per-coord
 // `D'²` weighting. Same structure as the 4-bit weighted kernel: i16
-// weights (i16-capped from u16 storage) → `madd_epi16` pair-sum →
-// widen to i64 for accumulation.
+// weights → `madd_epi16` pair-sum → widen to i64 for accumulation.
 //
 // Per-coord product bound (2-bit codebook):
 //   |c_a · c_b · w| ≤ 127·127·32 766 ≈ 5.28e8
@@ -411,7 +410,7 @@ pub unsafe fn score_2bit_internal_avx2(a: &[u8], b: &[u8]) -> f32 {
 /// # Safety
 /// CPU must support `ssse3` and `sse4.1`.
 #[target_feature(enable = "sse4.1,ssse3")]
-pub unsafe fn score_2bit_internal_weighted_sse(a: &[u8], b: &[u8], weights: &[u16]) -> i64 {
+pub unsafe fn score_2bit_internal_weighted_sse(a: &[u8], b: &[u8], weights: &[i16]) -> i64 {
     use core::arch::x86_64::*;
 
     assert_eq!(
@@ -482,7 +481,7 @@ pub unsafe fn score_2bit_internal_weighted_sse(a: &[u8], b: &[u8], weights: &[u1
 /// # Safety
 /// CPU must support `avx2`, `ssse3`, `sse4.1`.
 #[target_feature(enable = "avx2,sse4.1,ssse3")]
-pub unsafe fn score_2bit_internal_weighted_avx2(a: &[u8], b: &[u8], weights: &[u16]) -> i64 {
+pub unsafe fn score_2bit_internal_weighted_avx2(a: &[u8], b: &[u8], weights: &[i16]) -> i64 {
     use core::arch::x86_64::*;
 
     assert_eq!(a.len(), b.len());
@@ -667,12 +666,12 @@ mod tests {
     };
     use super::*;
 
-    /// Build deterministic i16-capped weights of length `4 · vec_bytes` for
-    /// parity tests of the 2-bit weighted kernel.
-    fn random_weights(rng: &mut StdRng, vec_bytes: usize) -> Vec<u16> {
+    /// Build deterministic non-negative i16 weights of length `4 · vec_bytes`
+    /// for parity tests of the 2-bit weighted kernel.
+    fn random_weights(rng: &mut StdRng, vec_bytes: usize) -> Vec<i16> {
         use rand::RngExt;
         (0..4 * vec_bytes)
-            .map(|_| rng.random_range(0..=i16::MAX as u16))
+            .map(|_| rng.random_range(0..=i16::MAX))
             .collect()
     }
 
@@ -856,8 +855,8 @@ mod tests {
         let indices: Vec<u8> = vec![3; dim];
         let vec_a = pack_codes(&indices, 2);
         let vec_b = pack_codes(&indices, 2);
-        let max_weight: u16 = i16::MAX as u16;
-        let weights: Vec<u16> = vec![max_weight; dim];
+        let max_weight: i16 = i16::MAX;
+        let weights: Vec<i16> = vec![max_weight; dim];
 
         let scalar = score_2bit_internal_weighted_scalar(&vec_a, &vec_b, &weights);
