+++ lib/quantization/src/turboquant/simd/query4bit/x64.rs
@@ -455,9 +455,10 @@ pub unsafe fn score_4bit_internal_avx512_vnni(a: &[u8], b: &[u8]) -> f32 {
 
 // ------------------------------------------------------------------
 // score_4bit_internal_weighted — TQ+ symmetric path with per-coord
-// `D'²` weighting. `weights[i]` is u16-quantized but i16-capped (see
-// `ErrorCorrection::d_prime_sq_u16` doc), so we treat the loaded bytes
-// as i16 and use the very efficient `madd_epi16` pair-sum.
+// `D'²` weighting. `weights[i]` is `i16` (non-negative, capped at
+// `i16::MAX − 1` — see `ErrorCorrection::d_prime_sq_i16` doc), which
+// matches the SIMD load directly and feeds the very efficient
+// `madd_epi16` pair-sum.
 //
 // Per-coord product bound:
 //   |c_a · c_b · w| ≤ 128·128·32 766 ≈ 5.37e8
@@ -473,7 +474,7 @@ pub unsafe fn score_4bit_internal_avx512_vnni(a: &[u8], b: &[u8]) -> f32 {
 /// # Safety
 /// CPU must support `ssse3` and `sse4.1`.
 #[target_feature(enable = "sse4.1,ssse3")]
-pub unsafe fn score_4bit_internal_weighted_sse(a: &[u8], b: &[u8], weights: &[u16]) -> i64 {
+pub unsafe fn score_4bit_internal_weighted_sse(a: &[u8], b: &[u8], weights: &[i16]) -> i64 {
     use core::arch::x86_64::*;
 
     assert_eq!(
@@ -566,7 +567,7 @@ pub unsafe fn score_4bit_internal_weighted_sse(a: &[u8], b: &[u8], weights: &[u1
 /// # Safety
 /// CPU must support `avx2`.
 #[target_feature(enable = "avx2")]
-pub unsafe fn score_4bit_internal_weighted_avx2(a: &[u8], b: &[u8], weights: &[u16]) -> i64 {
+pub unsafe fn score_4bit_internal_weighted_avx2(a: &[u8], b: &[u8], weights: &[i16]) -> i64 {
     use core::arch::x86_64::*;
 
     assert_eq!(
@@ -672,13 +673,12 @@ mod tests {
         score_4bit_internal_weighted_avx2, score_4bit_internal_weighted_sse,
     };
 
-    /// Build deterministic i16-capped weights of length `2 · vec_bytes` for
-    /// parity tests of the weighted kernels.  Same shape as the SIMD contract
-    /// (signed-i16 reinterpretation of u16 storage).
-    fn random_weights(rng: &mut StdRng, vec_bytes: usize) -> Vec<u16> {
+    /// Build deterministic non-negative i16 weights of length `2 · vec_bytes`
+    /// for parity tests of the weighted kernels.
+    fn random_weights(rng: &mut StdRng, vec_bytes: usize) -> Vec<i16> {
         use rand::RngExt;
         (0..2 * vec_bytes)
-            .map(|_| rng.random_range(0..=i16::MAX as u16))
+            .map(|_| rng.random_range(0..=i16::MAX))
             .collect()
     }
 
@@ -875,8 +875,8 @@ mod tests {
         let indices: Vec<u8> = vec![15; dim];
         let vec_a = pack_codes(&indices, 4);
         let vec_b = pack_codes(&indices, 4);
-        let max_weight: u16 = i16::MAX as u16;
-        let weights: Vec<u16> = vec![max_weight; dim];
+        let max_weight: i16 = i16::MAX;
+        let weights: Vec<i16> = vec![max_weight; dim];
 
         let scalar = score_4bit_internal_weighted_scalar(&vec_a, &vec_b, &weights);
